requirements.txt
# pip install -r requirements.txt
# For CUDA 11.8:
#   pip install torch==2.7.1+cu118 torchvision==0.22.1+cu118 --index-url https://download.pytorch.org/whl/cu118
# then:
#   pip install -r requirements.txt

segmentation-models-pytorch>=0.3.4
timm>=1.0.0
albumentations>=1.4.0
Pillow>=10.0
numpy>=2.1
scipy>=1.13
scikit-learn>=1.5
pandas>=2.2
matplotlib>=3.8
tqdm>=4.66
requests>=2.31
